fix video player: do not revalidate data after fetching vid url once
